main: don't set an upstream commit for a revert
We incorrectly mark a revert as an upstream commit when it really isn't
one (the revert is, or should, always be "local to that branch".)  So
don't allow a record to have both a revert and an upstream id.

This should help out in tracking reverts better.  We can now properly
query them from the database.  Tweaks to helper-tools to use this are
needed, but are underway...

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>