db: add indexes for several fields
Add some indexes for some of the fields in the main commits table.  This
increases the database size on disk to be almost double, BUT at runtime
it makes the generation of the CVE data 5x faster, making this very
worth it.

Note, maybe not all indexes are needed just now, but it can't hurt to
have as many different tools query different things.

Taken from a larger patch from Sasha

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>